refactor: move all double data type to float (#98)
Move all double data type to float.

This is not really necessary, because soon it should allow both float and double, but we are doing it now in order to identify problems with incompatibility and we need to make the code more flexible.

+++ src/lexer.cpp
@@ -9,7 +9,7 @@ SourceLocation Lexer::cur_loc;
 // Filled in if tok_identifier
 std::string Lexer::identifier_str = "<NOT DEFINED>";
 // Filled in if tok_float_literal
-double Lexer::num_float;
+float Lexer::num_float;
 SourceLocation Lexer::lex_loc;
 int Lexer::cur_tok = tok_not_initialized;
